Leetcode 1727 & 3070

+++ go/medium/3296.Minimum-Number-of-Seconds-to-Make-Mountain-Height-Zero/3296.go
@@ -34,5 +34,5 @@ func minNumberOfSeconds(mountainHeight int, workerTimes []int) int64 {
 			lo = mid + 1
 		}
 	}
-	return lo
+	return int64(lo)
 }
